Routing added
Application was separated to two screens - one, which displays tasks to
be done with add-task component and save button, second for tasks-done
list.

Routing was implemented in app.routing.model.ts. It was imported in
app.module.ts.

Components had small refactor to suit new way of accessing the
aplication.